+++ R/plot_light_effects.R
@@ -1,13 +1,13 @@
 #' Visualize Different Types of Profiles Together
 #'
-#' Visualizes response-, prediction-, and partial dependence profiles of a (multi-)flashlight with respect to a covariable \code{v}. Different flashlights or a single flashlight with one "by" variable are separated by a facet wrap.
+#' Visualizes response-, prediction-, partial dependence, and/or ALE profiles of a (multi-)flashlight with respect to a covariable \code{v}. Different flashlights or a single flashlight with one "by" variable are separated by a facet wrap.
 #'
 #' @import ggplot2
 #' @importFrom stats reformulate
 #' @importFrom dplyr semi_join bind_rows
 #' @method plot light_effects
 #' @param x An object of class \code{light_effects}.
-#' @param use A vector with elements in "response", "predicted" and "pd" (partial dependence) with elements to show. Defaults to all.
+#' @param use A vector with elements in "response", "predicted", "pd" (partial dependence) and "ale" with elements to show. Defaults to all except "ale". Use "all" for all.
 #' @param zero_counts Logical flag if 0 count levels should be shown on the x axis.
 #' @param size_factor Factor used to enlarge default \code{size} in \code{geom_point} and \code{geom_line}.
 #' @param facet_scales Scales argument passed to \code{facet_wrap}.
@@ -49,6 +49,9 @@ plot.light_effects <- function(x, use = c("response", "predicted", "pd"),
                                facet_nrow = 1L, rotate_x = TRUE, ...) {
   # Checks
   stopifnot(length(use) >= 1L)
+  if ("all" %in% use) {
+    use <- c("response", "predicted", "pd", "ale")
+  }
   nby <- length(x$by)
   multi <- is.light_effects_multi(x)
   if (nby + multi > 1L) {