+++ packaging.R
@@ -22,9 +22,9 @@ create_package(
   fields = list(
     Title = "Shed Light on Black Box Machine Learning Models",
     Type = "Package",
-    Version = "0.3.1",
+    Version = "0.4.0",
     Date = Sys.Date(),
-    Description = "Shed light on black box machine learning models by the help of model performance, permutation variable importance (Fisher et al. (2018) <arxiv:1801.01489>), ICE profiles, partial dependence (Friedman J. H. (2001) <doi:10.1214/aos/1013203451>), accumulated local effects (Apley D. W. (2016) <arXiv:1612.08468>), further effects plots, and variable contribution breakdown for single observations (Gosiewska and Biecek (2019) <arxiv:1903.11420>). All tools are implemented to work with case weights and allow for stratified analysis. Furthermore, multiple flashlights can be combined and analyzed together.",
+    Description = "Shed light on black box machine learning models by the help of model performance, permutation variable importance (Fisher et al. (2018) <arxiv:1801.01489>), ICE profiles, partial dependence (Friedman J. H. (2001) <doi:10.1214/aos/1013203451>), accumulated local effects (Apley D. W. (2016) <arXiv:1612.08468>), further effects plots, interaction strength, and variable contribution breakdown for single observations (Gosiewska and Biecek (2019) <arxiv:1903.11420>). All tools are implemented to work with case weights and allow for stratified analysis. Furthermore, multiple flashlights can be combined and analyzed together.",
     `Authors@R` = "person('Michael', 'Mayer', email = 'mayermichael79@gmail.com', role = c('aut', 'cre', 'cph'))",
     Depends = "R (>= 3.5.0)",
     VignetteBuilder = "knitr",
@@ -42,7 +42,7 @@ use_package("tidyr", "Imports")
 use_package("rlang", "Imports")
 use_package("ggplot2", "Imports")
 use_package("ggpubr", "Imports")
-use_package("MetricsWeighted", "Imports", min_version = "0.2.0")
+use_package("MetricsWeighted", "Imports", min_version = "0.3.0")
 
 # Suggests
 use_package("knitr", "Suggests")